+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Fixes:
 
 ### master (unreleased)
 
+Misc:
+
+- [#123](https://github.com/graphiti-api/graphiti/pull/123) Throw
+  better error when polymorphic type not found.
+
 <!-- ### [version (YYYY-MM-DD)](diff_link) -->
 <!-- Breaking changes:-->
 <!-- Features:-->